requirements.txt
pandas
scikit-learn
dvc[s3]
boto3
fastapi
uvicorn[standard]
python-multipart
PyYAML
